+++ src/configs/hotspot-detection.config.ts
@@ -1,0 +1,5 @@
+import { getEnv } from '@/shared/utils/env.util';
+
+export const hotspotDetectionConfig = {
+    url: getEnv("HOTSPOT_DETECTION_URL")
+} as const;