+++ src/configs/index.ts
@@ -1,5 +1,6 @@
 import { appConfig } from '@/configs/app.config';
 import { dbConfig } from '@/configs/db.config';
+import { hotspotDetectionConfig } from '@/configs/hotspot-detection.config';
 import { mediamtxConfig } from '@/configs/mediamtx.config';
 import { mqttConfig } from '@/configs/mqtt.config';
 import { redisConfig } from '@/configs/redis.config';
@@ -10,4 +11,5 @@ export const config = {
     redis: redisConfig,
     mqtt: mqttConfig,
     mediamtx: mediamtxConfig,
+    hotspotDetection: hotspotDetectionConfig,
 } as const;