+++ src/configs/mediamtx.config.ts
@@ -1,5 +1,6 @@
 import { getEnv } from '@/shared/utils/env.util';
 
 export const mediamtxConfig = {
-    url: getEnv("MEDIA_MTX_URL")
+    api: getEnv('MEDIA_MTX_URL'),
+    rtsp: getEnv('MEDIA_MTX_RTSP'),
 } as const;